refactoring: ErrorResponse 패키지 위치 수정 및 message 필드 제거
