+++ src/main/java/com/example/springboard/common/handler/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.example.springboard.common.handler;
 
-import com.example.springboard.common.ErrorResponse;
 import com.example.springboard.common.exception.GlobalException;
+import com.example.springboard.dto.response.ErrorResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
@@ -16,8 +16,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(GlobalException.class)
     public ResponseEntity<ErrorResponse> handleException(GlobalException exception) {
         // todo logger 처리
-        ErrorResponse errorResponse = new ErrorResponse(exception.getErrorType(),
-            exception.getErrorType().getMessage());
+        ErrorResponse errorResponse = new ErrorResponse(exception.getErrorType());
         if (exception.getErrorType().getCode() >= 500_000) {
             return ResponseEntity.internalServerError().body(errorResponse);
         } else if (exception.getErrorType().getCode() >= 400_000) {
