+++ src/main/java/com/example/springboard/util/enums/ErrorTypeWithRequest.java
@@ -57,4 +57,8 @@ public enum ErrorTypeWithRequest {
     ErrorTypeWithRequest(int code) {
         this.code = code;
     }
+
+    public int getCode() {
+        return code;
+    }
 }
